🔨 Fix: MyPageFragment 두 개인 버그 수정
