+++ app/src/main/res/layout/member_fragment_my_page.xml
@@ -2,8 +2,7 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
-             android:layout_height="match_parent"
-             tools:context=".MyPageFragment">
+             android:layout_height="match_parent">
 
     <TextView
             android:layout_width="wrap_content"
