+++ app/src/main/res/navigation/nav_graph.xml
@@ -64,7 +64,7 @@
 
     <fragment
             android:id="@+id/myPageFragment"
-            android:name="com.otclub.humate.MyPageFragment"
+            android:name="com.otclub.humate.member.fragment.MyPageFragment"
             android:label="fragment_my_page"
             tools:layout="@layout/member_fragment_my_page"/>
     <fragment
